prjconf: Allow for configs to have RPM related inc file extension
The syntax inside the project configuration is not XML but RPM-like.
Instead of the xml extension the inc extension makes more sense here,
support both file extensions.
Inc was the closest file extension to this use case,
as the configuration fragments are similarly used
as in RPM packages %include files.

[prjconf] Allow for configs to have RPM related inc file extension. JB#51936

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>